refactor: Decouple core from environment
This commit refactors the `core` package to remove direct access to
the file system and environment variables. The `cli` package is now
responsible for reading the configuration from the environment and
passing it to the `core` package. This makes the `core` package
more portable and easier to test.
